feat: :zap: add tehem swicht html
